+++ src/interfaces/IEmailAuth.sol
@@ -4,10 +4,7 @@ pragma solidity ^0.8.25;
 interface IEmailAuth {
     function updateDKIMRegistry(address dkimRegistryAddr) external;
     function updateVerifier(address verifierAddr) external;
-    function updateSubjectTemplate(
-        uint templateId,
-        string[] memory subjectTemplate
-    ) external;
-    function deleteSubjectTemplate(uint templateId) external;
+    function updateSubjectTemplate(uint256 templateId, string[] memory subjectTemplate) external;
+    function deleteSubjectTemplate(uint256 templateId) external;
     function setTimestampCheckEnabled(bool enabled) external;
 }
